+++ baseapp/circuit.go
@@ -1,8 +1,10 @@
 package baseapp
 
-import "context"
+import (
+	sdk "github.com/cosmos/cosmos-sdk/types"
+)
 
 // CircuitBreaker is an interface that defines the methods for a circuit breaker.
 type CircuitBreaker interface {
-	IsAllowed(ctx context.Context, typeURL string) (bool, error)
+	IsAllowed(ctx sdk.Context, typeURL string) bool
 }